chore!: convert to a pure esm module
BREAKING CHANGE: convert npm-check
to a pure esm module as many of its
dependencies have taken this approach

see https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c
