+++ README.md
@@ -176,7 +176,7 @@ Enable or disable the spinner. Useful for terminals that don't support them. Aut
 The API is here in case you want to wrap this with your CI toolset.
 
 ```js
-const npmCheck = require('npm-check');
+import npmCheck from 'npm-check';
 
 npmCheck(options)
   .then(currentState => console.log(currentState.get('packages')));
