+++ lib/in/best-guess-homepage.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const gitUrl = require('giturl');
+import gitUrl from "giturl";
 
 function bestGuessHomepage(data) {
     if (!data) {
@@ -14,4 +14,4 @@ function bestGuessHomepage(data) {
         packageDataForLatest.repository && packageDataForLatest.repository.url && gitUrl.parse(packageDataForLatest.repository.url.trim());
 }
 
-module.exports = bestGuessHomepage;
+export default bestGuessHomepage;
