+++ lib/in/create-package-summary.js
@@ -1,20 +1,20 @@
 'use strict';
 
-const readPackageJson = require('./read-package-json');
-const getLatestFromRegistry = require('./get-latest-from-registry');
-const findModulePath = require('./find-module-path');
-const _ = require('lodash');
-const semverDiff = require('semver-diff');
-const pathExists = require('path-exists');
-const path = require('path');
-const semver = require('semver');
-const minimatch = require('minimatch');
+import readPackageJson from "./read-package-json.js";
+import getLatestFromRegistry from "./get-latest-from-registry.js";
+import findModulePath from "./find-module-path.js";
+import _ from "lodash";
+import semverDiff from "semver-diff";
+import { pathExistsSync } from "path-exists";
+import path from "path";
+import semver from "semver";
+import minimatch from "minimatch";
 
 function createPackageSummary(moduleName, currentState) {
     const cwdPackageJson = currentState.get('cwdPackageJson');
 
     const modulePath = findModulePath(moduleName, currentState);
-    const packageIsInstalled = pathExists.sync(modulePath);
+    const packageIsInstalled = pathExistsSync(modulePath);
     const modulePackageJson = readPackageJson(path.join(modulePath, 'package.json'));
 
     // Ignore private packages
@@ -110,4 +110,4 @@ function createPackageSummary(moduleName, currentState) {
         });
 }
 
-module.exports = createPackageSummary;
+export default createPackageSummary;
