+++ lib/in/find-module-path.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const Module = require('module');
-const path = require('path');
-const pathExists = require('path-exists');
+import Module from "module";
+import path from "path";
+import { pathExistsSync } from "path-exists";
 
 /**
  * Searches the directory hierarchy to return the path to the requested node module.
@@ -21,10 +21,10 @@ function findModulePath(moduleName, currentState) {
     // We can't use require.resolve because it fails if the module doesn't have an entry point.
     const nodeModulesPaths = Module._nodeModulePaths(cwd);
     const possibleModulePaths = nodeModulesPaths.map(x => path.join(x, moduleName));
-    const modulePath = possibleModulePaths.find(pathExists.sync);
+    const modulePath = possibleModulePaths.find(pathExistsSync);
 
     // if no existing path was found, return the first tried path anyway
     return modulePath || path.join(cwd, moduleName);
 }
 
-module.exports = findModulePath;
+export default findModulePath;
