+++ lib/in/get-installed-packages.js
@@ -1,10 +1,10 @@
 'use strict';
-const _ = require('lodash');
-const globby = require('globby');
-const readPackageJson = require('./read-package-json');
-const path = require('path');
+import _ from "lodash";
+import { globby } from "globby";
+import readPackageJson from "./read-package-json.js";
+import path from "path";
 
-module.exports = function (cwd) {
+export default function (cwd) {
     const GLOBBY_PACKAGE_JSON = '{*/package.json,@*/*/package.json}';
     const installedPackages = globby.sync(GLOBBY_PACKAGE_JSON, {cwd: cwd});
 
@@ -15,4 +15,4 @@ module.exports = function (cwd) {
         })
         .fromPairs()
         .valueOf();
-};
+}
