+++ lib/in/get-latest-from-registry.js
@@ -1,11 +1,14 @@
 'use strict';
 
-const _ = require('lodash');
-const bestGuessHomepage = require('./best-guess-homepage');
-const semver = require('semver');
-const packageJson = require('package-json');
-const cpuCount = require('os').cpus().length;
-const throat = require('throat')(cpuCount);
+import _ from "lodash";
+import bestGuessHomepage from "./best-guess-homepage.js";
+import semver from "semver";
+import packageJson from "package-json";
+import {cpus} from "os";
+import throat0 from "throat";
+
+const cpuCount = cpus().length;
+const throat = throat0(cpuCount);
 
 function getNpmInfo(packageName) {
     return throat(() => packageJson(packageName, { fullMetadata: true, allVersions: true }))
@@ -37,4 +40,4 @@ function getNpmInfo(packageName) {
         });
 }
 
-module.exports = getNpmInfo;
+export default getNpmInfo;
