+++ lib/in/get-unused-packages.js
@@ -1,9 +1,10 @@
 'use strict';
 
-const depcheck = require('depcheck');
-const ora = require('ora');
-const _ = require('lodash');
-const { rcFile } = require('rc-config-loader');
+import depcheck from "depcheck";
+import ora from "ora";
+import _ from "lodash";
+import {rcFile} from "rc-config-loader";
+
 
 function skipUnused(currentState) {
     return currentState.get('skipUnused') ||        // manual option to ignore this
@@ -95,4 +96,4 @@ function checkUnused(currentState) {
     });
 }
 
-module.exports = checkUnused;
+export default checkUnused;
