+++ lib/in/index.js
@@ -1,11 +1,11 @@
 'use strict';
-const co = require('co');
-const extend = require('xtend');
-const ora = require('ora');
-const getUnusedPackages = require('./get-unused-packages');
-const createPackageSummary = require('./create-package-summary');
+import co from "co";
+import extend from "xtend";
+import ora from "ora";
+import getUnusedPackages from "./get-unused-packages.js";
+import createPackageSummary from "./create-package-summary.js";
 
-module.exports = function (currentState) {
+export default function (currentState) {
     return co(function *() {
         yield getUnusedPackages(currentState);
 
@@ -43,4 +43,4 @@ module.exports = function (currentState) {
         spinner.stop();
         return currentState;
     });
-};
+}
