+++ lib/in/read-package-json.js
@@ -1,12 +1,16 @@
 'use strict';
 
-const extend = require('xtend');
+import { readFileSync } from 'fs';
+import extend from "xtend";
 
 function readPackageJson(filename) {
     let pkg;
     let error;
     try {
-        pkg = require(filename);
+        pkg = JSON.parse(readFileSync(
+            new URL(filename, import.meta.url),
+            { encoding: 'utf8' },
+        ));
     } catch (e) {
         if (e.code === 'MODULE_NOT_FOUND') {
             error = new Error(`A package.json was not found at ${filename}`);
@@ -17,4 +21,4 @@ function readPackageJson(filename) {
     return extend({devDependencies: {}, dependencies: {}, error: error}, pkg)
 }
 
-module.exports = readPackageJson;
+export default readPackageJson;
