+++ lib/index.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const npmCheck = require('./in');
-const createState = require('./state/state');
+import npmCheck from "./in/index.js";
+import createState from "./state/state.js";
 
 async function init(userOptions) {
     return npmCheck(await createState(userOptions));
 }
 
-module.exports = init;
+export default init;
