+++ lib/state/debug.js
@@ -1,5 +1,5 @@
 'use strict';
-const chalk = require('chalk');
+import chalk from "chalk";
 
 function debug() {
     console.log(chalk.green('[npm-check] debug'));
@@ -7,4 +7,4 @@ function debug() {
     console.log(`${chalk.green('===============================')}`);
 }
 
-module.exports = debug;
+export default debug;
