+++ lib/state/init.js
@@ -1,12 +1,12 @@
 'use strict';
-const _ = require('lodash');
-const path = require('path');
-const globalModulesPath = require('global-modules');
-const readPackageJson = require('../in/read-package-json');
-const globalPackages = require('../in/get-installed-packages');
-const emoji = require('../out/emoji');
-const fs = require('fs');
-const chalk = require('chalk');
+import _ from "lodash";
+import path from "path";
+import globalModulesPath from "global-modules";
+import readPackageJson from "../in/read-package-json.js";
+import globalPackages from "../in/get-installed-packages.js";
+import emoji, { enabled as emojiEnabled} from "../out/emoji.js";
+import fs from "fs";
+import chalk from "chalk";
 
 function init(currentState, userOptions) {
     return new Promise((resolve, reject) => {
@@ -39,7 +39,7 @@ function init(currentState, userOptions) {
             currentState.set('cwd', cwd);
         }
 
-        emoji.enabled(currentState.get('emoji'));
+        emojiEnabled(currentState.get('emoji'));
 
         if (currentState.get('cwdPackageJson').error) {
             return reject(currentState.get('cwdPackageJson').error);
@@ -49,4 +49,4 @@ function init(currentState, userOptions) {
     });
 }
 
-module.exports = init;
+export default init;
