+++ lib/state/state.js
@@ -1,7 +1,7 @@
 'use strict';
-const extend = require('xtend');
-const init = require('./init');
-const debug = require('./debug');
+import extend from "xtend";
+import init from "./init.js";
+import debug from "./debug.js";
 
 const defaultOptions = {
     update: false,
@@ -74,4 +74,5 @@ function state(userOptions) {
 
     return init(currentState, userOptions);
 }
-module.exports = state;
+
+export default state;
