+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "npm-check",
   "version": "6.0.1",
+  "type": "module",
   "description": "Check for outdated, incorrect, and unused dependencies.",
   "main": "lib",
   "engines": {
